+++ GPy/kern/src/psi_comp/ssrbf_psi_gpucomp.py
@@ -314,32 +314,32 @@ class PSICOMP_SSRBF_GPU(PSICOMP_RBF):
         import pycuda.gpuarray as gpuarray
         if self.gpuCache == None:
             self.gpuCache = {
-                             'l_gpu'                :gpuarray.empty((Q,),np.float64,order='F'),
-                             'Z_gpu'                :gpuarray.empty((M,Q),np.float64,order='F'),
-                             'mu_gpu'               :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'S_gpu'                :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'gamma_gpu'            :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'psi1_gpu'             :gpuarray.empty((N,M),np.float64,order='F'),
-                             'psi2_gpu'             :gpuarray.empty((M,M),np.float64,order='F'),
-                             'psi2n_gpu'            :gpuarray.empty((N,M,M),np.float64,order='F'),
-                             'dL_dpsi1_gpu'         :gpuarray.empty((N,M),np.float64,order='F'),
-                             'dL_dpsi2_gpu'         :gpuarray.empty((M,M),np.float64,order='F'),
-                             'log_denom1_gpu'       :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'log_denom2_gpu'       :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'log_gamma_gpu'        :gpuarray.empty((N,Q),np.float64,order='F'),
-                             'log_gamma1_gpu'       :gpuarray.empty((N,Q),np.float64,order='F'),
+                             'l_gpu'                :gpuarray.empty((Q,),float,order='F'),
+                             'Z_gpu'                :gpuarray.empty((M,Q),float,order='F'),
+                             'mu_gpu'               :gpuarray.empty((N,Q),float,order='F'),
+                             'S_gpu'                :gpuarray.empty((N,Q),float,order='F'),
+                             'gamma_gpu'            :gpuarray.empty((N,Q),float,order='F'),
+                             'psi1_gpu'             :gpuarray.empty((N,M),float,order='F'),
+                             'psi2_gpu'             :gpuarray.empty((M,M),float,order='F'),
+                             'psi2n_gpu'            :gpuarray.empty((N,M,M),float,order='F'),
+                             'dL_dpsi1_gpu'         :gpuarray.empty((N,M),float,order='F'),
+                             'dL_dpsi2_gpu'         :gpuarray.empty((M,M),float,order='F'),
+                             'log_denom1_gpu'       :gpuarray.empty((N,Q),float,order='F'),
+                             'log_denom2_gpu'       :gpuarray.empty((N,Q),float,order='F'),
+                             'log_gamma_gpu'        :gpuarray.empty((N,Q),float,order='F'),
+                             'log_gamma1_gpu'       :gpuarray.empty((N,Q),float,order='F'),
                              # derivatives
-                             'dvar_gpu'             :gpuarray.empty((self.blocknum,),np.float64, order='F'),
-                             'dl_gpu'               :gpuarray.empty((Q,self.blocknum),np.float64, order='F'),
-                             'dZ_gpu'               :gpuarray.empty((M,Q),np.float64, order='F'),
-                             'dmu_gpu'              :gpuarray.empty((N,Q,self.blocknum),np.float64, order='F'),
-                             'dS_gpu'               :gpuarray.empty((N,Q,self.blocknum),np.float64, order='F'),
-                             'dgamma_gpu'           :gpuarray.empty((N,Q,self.blocknum),np.float64, order='F'),
+                             'dvar_gpu'             :gpuarray.empty((self.blocknum,),float, order='F'),
+                             'dl_gpu'               :gpuarray.empty((Q,self.blocknum),float, order='F'),
+                             'dZ_gpu'               :gpuarray.empty((M,Q),float, order='F'),
+                             'dmu_gpu'              :gpuarray.empty((N,Q,self.blocknum),float, order='F'),
+                             'dS_gpu'               :gpuarray.empty((N,Q,self.blocknum),float, order='F'),
+                             'dgamma_gpu'           :gpuarray.empty((N,Q,self.blocknum),float, order='F'),
                              # grad
-                             'grad_l_gpu'               :gpuarray.empty((Q,),np.float64, order='F'),
-                             'grad_mu_gpu'              :gpuarray.empty((N,Q,),np.float64, order='F'),
-                             'grad_S_gpu'               :gpuarray.empty((N,Q,),np.float64, order='F'),
-                             'grad_gamma_gpu'           :gpuarray.empty((N,Q,),np.float64, order='F'),
+                             'grad_l_gpu'               :gpuarray.empty((Q,),float, order='F'),
+                             'grad_mu_gpu'              :gpuarray.empty((N,Q,),float, order='F'),
+                             'grad_S_gpu'               :gpuarray.empty((N,Q,),float, order='F'),
+                             'grad_gamma_gpu'           :gpuarray.empty((N,Q,),float, order='F'),
                              }
         else:
             assert N==self.gpuCache['mu_gpu'].shape[0]
@@ -399,8 +399,8 @@ class PSICOMP_SSRBF_GPU(PSICOMP_RBF):
 
         psi0 = np.empty((N,))
         psi0[:] = variance
-        self.g_psi1computations.prepared_call((self.blocknum,1),(self.threadnum,1,1),psi1_gpu.gpudata, log_denom1_gpu.gpudata, log_gamma_gpu.gpudata, log_gamma1_gpu.gpudata, np.float64(variance),l_gpu.gpudata,Z_gpu.gpudata,mu_gpu.gpudata,S_gpu.gpudata, np.int32(N), np.int32(M), np.int32(Q))
-        self.g_psi2computations.prepared_call((self.blocknum,1),(self.threadnum,1,1),psi2_gpu.gpudata, psi2n_gpu.gpudata, log_denom2_gpu.gpudata, log_gamma_gpu.gpudata, log_gamma1_gpu.gpudata, np.float64(variance),l_gpu.gpudata,Z_gpu.gpudata,mu_gpu.gpudata,S_gpu.gpudata, np.int32(N), np.int32(M), np.int32(Q))
+        self.g_psi1computations.prepared_call((self.blocknum,1),(self.threadnum,1,1),psi1_gpu.gpudata, log_denom1_gpu.gpudata, log_gamma_gpu.gpudata, log_gamma1_gpu.gpudata, float(variance),l_gpu.gpudata,Z_gpu.gpudata,mu_gpu.gpudata,S_gpu.gpudata, np.int32(N), np.int32(M), np.int32(Q))
+        self.g_psi2computations.prepared_call((self.blocknum,1),(self.threadnum,1,1),psi2_gpu.gpudata, psi2n_gpu.gpudata, log_denom2_gpu.gpudata, log_gamma_gpu.gpudata, log_gamma1_gpu.gpudata, float(variance),l_gpu.gpudata,Z_gpu.gpudata,mu_gpu.gpudata,S_gpu.gpudata, np.int32(N), np.int32(M), np.int32(Q))
         
         if self.GPU_direct:
             return psi0, psi1_gpu, psi2_gpu
@@ -449,12 +449,12 @@ class PSICOMP_SSRBF_GPU(PSICOMP_RBF):
 
         self.reset_derivative()
         # t=self.g_psi1compDer(dvar_gpu,dl_gpu,dZ_gpu,dmu_gpu,dS_gpu,dL_dpsi1_gpu,psi1_gpu,
-        #                      np.float64(variance),l_gpu,Z_gpu,mu_gpu,S_gpu, np.int32(N),
+        #                      float(variance),l_gpu,Z_gpu,mu_gpu,S_gpu, np.int32(N),
         #                      np.int32(M), np.int32(Q), block=(self.threadnum,1,1),
         #                      grid=(self.blocknum,1),time_kernel=True)
         # print 'g_psi1compDer '+str(t)
         # t=self.g_psi2compDer(dvar_gpu,dl_gpu,dZ_gpu,dmu_gpu,dS_gpu,dL_dpsi2_gpu,psi2n_gpu,
-        #                      np.float64(variance),l_gpu,Z_gpu,mu_gpu,S_gpu, np.int32(N),
+        #                      float(variance),l_gpu,Z_gpu,mu_gpu,S_gpu, np.int32(N),
         #                      np.int32(M), np.int32(Q), block=(self.threadnum,1,1),
         #                      grid=(self.blocknum,1),time_kernel=True)
         # print 'g_psi2compDer '+str(t)
@@ -463,7 +463,7 @@ class PSICOMP_SSRBF_GPU(PSICOMP_RBF):
                                          dmu_gpu.gpudata, dS_gpu.gpudata, dgamma_gpu.gpudata,
                                          dL_dpsi1_gpu.gpudata, psi1_gpu.gpudata,
                                          log_denom1_gpu.gpudata, log_gamma_gpu.gpudata,
-                                         log_gamma1_gpu.gpudata, np.float64(variance),
+                                         log_gamma1_gpu.gpudata, float(variance),
                                          l_gpu.gpudata, Z_gpu.gpudata, mu_gpu.gpudata,
                                          S_gpu.gpudata, gamma_gpu.gpudata, np.int32(N),
                                          np.int32(M), np.int32(Q))
@@ -472,7 +472,7 @@ class PSICOMP_SSRBF_GPU(PSICOMP_RBF):
                                          dmu_gpu.gpudata, dS_gpu.gpudata, dgamma_gpu.gpudata,
                                          dL_dpsi2_gpu.gpudata, psi2n_gpu.gpudata,
                                          log_denom2_gpu.gpudata, log_gamma_gpu.gpudata,
-                                         log_gamma1_gpu.gpudata, np.float64(variance),
+                                         log_gamma1_gpu.gpudata, float(variance),
                                          l_gpu.gpudata, Z_gpu.gpudata, mu_gpu.gpudata,
                                          S_gpu.gpudata, gamma_gpu.gpudata, np.int32(N),
                                          np.int32(M), np.int32(Q))
