+++ GPy/util/misc.py
@@ -5,7 +5,7 @@ import numpy as np
 from scipy.special import cbrt
 from .config import *
 
-_lim_val = np.finfo(np.float64).max
+_lim_val = np.finfo(float).max
 _lim_val_exp = np.log(_lim_val)
 _lim_val_square = np.sqrt(_lim_val)
 #_lim_val_cube = cbrt(_lim_val)
